Migrate dependencies and import functions needed in webACS.js (and their dependencies, unfinsihed)
