+++ assets/resources/defaultComponentCollection.abd
