+++ package.json
@@ -18,6 +18,7 @@
     "@babel/preset-env": "^7.3.4",
     "@babel/register": "^7.0.0",
     "babel-loader": "^8.0.5",
+    "copy-webpack-plugin": "^5.0.1",
     "cross-env": "^5.2.0",
     "css-loader": "^2.1.1",
     "favicons-webpack-plugin": "^0.0.9",
