+++ src/model/clipBoard.js
@@ -26,7 +26,7 @@
  * limitations under the License.
  */
  
-ACS.clipBoard = function() {
+export default function() {
 
 // ***********************************************************************************************************************
 // ************************************************** private variables **************************************************
